+++ src/ray.rs
@@ -1,6 +1,6 @@
 use nalgebra::{Vector3, Point3};
 
 pub struct Ray {
-    pub origin: Point3<f32>,
-    pub direction: Vector3<f32>,
+    pub origin: Point3<f64>,
+    pub direction: Vector3<f64>,
 }
